fix: metadata object and rename Nip05 service file
