+++ lib/nostr_client.dart
@@ -9,6 +9,8 @@ export 'package:nostr_client/src/key_pair/generator/random_key_pair_generator.da
 export 'package:nostr_client/src/key_pair/key_pair.dart';
 export 'package:nostr_client/src/message.dart';
 export 'package:nostr_client/src/metadata.dart';
+export 'package:nostr_client/src/nip05.dart';
+export 'package:nostr_client/src/nip05_service.dart';
 export 'package:nostr_client/src/relay/relay.dart';
 export 'package:nostr_client/src/relay/relay_information_document.dart';
 export 'package:nostr_client/src/relay/relay_pool.dart';
